# internal/dto/response_dto.go
package dto

// SuccessResponse represents a standard success response
// swagger:model
type SuccessResponse struct {
	Status string      `json:"status"`
	Data   interface{} `json:"data"`
}

type ErrorResponse struct {
	Status  string `json:"status"`
	Message string `json:"message"`
}